Restructure the code Added flow-bin as a dev deps again
