+++ package.json
@@ -34,5 +34,8 @@
     "eslint-config-henribeck": "4.19.6",
     "flow-bin": "0.73.0",
     "rimraf": "2.6.2"
+  },
+  "peerDependencies": {
+    "flow-bin": "^0.65.0"
   }
 }
